tests/wpt/tests/navigation-api/navigate-event/navigate-anchor-cross-origin.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<a id="a"></a>
<script>
async_test(t => {
  navigation.onnavigate = t.step_func_done(e => {
    assert_equals(e.navigationType, "push");
    assert_true(e.cancelable);
    assert_false(e.canIntercept);
    assert_false(e.userInitiated);
    assert_false(e.hashChange);
    assert_equals(e.formData, null);
    assert_equals(e.downloadRequest, null);
    assert_equals(e.destination.url, a.href);
    assert_false(e.destination.sameDocument);
    assert_equals(e.destination.key, "");
    assert_equals(e.destination.id, "");
    assert_equals(e.destination.index, -1);
    assert_equals(e.sourceElement, document.getElementById("a"));
    e.preventDefault();
  });
  a.href = get_host_info().HTTPS_REMOTE_ORIGIN + "/does-not-exist.html";
  a.click();
}, "<a> cross-origin navigate event");
</script>
