LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-anchor-cross-origin-expected.txt


FAIL <a> cross-origin navigate event assert_equals: expected (object) Element node <a id="a" href="https://does-not-exist/foo.html"></a> but got (undefined) undefined
